телеграмм бот/requirements.txt
python-telegram-bot==20.3
pillow==10.1.0
opencv-python==4.8.1
numpy==1.24.3
scikit-image==0.22.0